Added custom response message for DELETE, POST; formatting

+++ src/main/java/com/example/demo/controller/CarController.java
@@ -1,6 +1,7 @@
 package com.example.demo.controller;
 
 import com.example.demo.assembler.CarResourceAssembler;
+import com.example.demo.customJson.CustomJSONResponse;
 import com.example.demo.entity.Car;
 import com.example.demo.errors.CarNotFoundException;
 import com.example.demo.repository.CarRepository;
@@ -8,6 +9,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.RepresentationModel;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -38,8 +40,7 @@ public class CarController {
 
   @GetMapping("/cars")
   public ResponseEntity<CollectionModel<EntityModel<Car>>> getCars() {
-    return ResponseEntity
-            .ok(assembler.toCollectionModel(carRepository.findAll()));
+    return ResponseEntity.ok(assembler.toCollectionModel(carRepository.findAll()));
   }
 
   @GetMapping("/cars/{id}")
@@ -54,16 +55,22 @@ public class CarController {
   @PostMapping("/cars")
   public ResponseEntity<EntityModel<Car>> addCar(@RequestBody Car car) {
     Car savedCar = carRepository.save(Objects.requireNonNull(car));
-    return ResponseEntity
-            .created(linkTo(methodOn(CarController.class)
-                             .getCar(car.getId()))
-                             .toUri())
+    return ResponseEntity.created(linkTo(methodOn(CarController.class).getCar(car.getId())).toUri())
             .body(assembler.toModel(savedCar));
   }
 
   @DeleteMapping("/cars/{id}")
   public ResponseEntity<?> deleteCar(@PathVariable Long id) {
-    carRepository.deleteById(id);
-    return ResponseEntity.ok("ID " + id + " deleted");
+    if (carRepository.existsById(id)) {
+      carRepository.deleteById(id);
+      return new ResponseEntity<>(
+              new CustomJSONResponse(
+                      String.valueOf(HttpStatus.ACCEPTED.value()),
+                      HttpStatus.ACCEPTED.getReasonPhrase(),
+                      "Car with id " + id + " has been deleted."),
+              HttpStatus.ACCEPTED);
+    } else {
+      throw new CarNotFoundException(id);
+    }
   }
 }
